🐛 Rename scheduler to avoid naming conflict
