+++ main.py
@@ -15,7 +15,7 @@ import new_xkcd
 import rand_frog
 import reddit
 import replace_vowels
-import scheduler
+import homemade_scheduler
 import scores
 import text_gen
 from telegram import Telegram, user_name
@@ -23,7 +23,7 @@ from wolfram_alpha import query_wa
 
 tg = Telegram(config.token)
 next_launch = None
-bot_scheduler = scheduler.Scheduler()
+bot_scheduler = homemade_scheduler.Scheduler()
 
 
 def main():
